Add Client Pulp service
This change adds playbooks and variables for managing a client Pulp service.
The repositories in the client Pulp service are synced with the release Pulp
service, and each have 3 distributions - development, staging and production.
Each repository has a variable which defines the version that will be synced
into the development distribution. A playbook is provided to automatically
update these versions to the latest available. Playbooks are provided to
promote development to staging and then production.