Workaroud Pulp multiple publications bug
Due to a bug in Pulp (https://github.com/pulp/pulp_rpm/issues/2311), we can
end up with multiple publications for a single repository version after
performing a sync. This breaks assumptions in the Pulp squeezer modules about
uniqueness, causing the rpm_publication module to fail, and breaking sync
jobs.

This patch adds a playbook that looks for repository versions with
multiple publications, and removes the most recent. The playbook is
added to the Github repo-sync workflow.