test: query package repository versions to sync dynamically
Currently, the set of package repository versions to sync from Ark to
test is defined via a group_vars file in the inventory.  In practice, we
always update the versions to the latest in Ark before syncing.

It has become clear that this repository is not the right place to store
'live' versions - that is stackhpc-kayobe-config.

This change avoids references to the group_vars versions when syncing,
and adds a new 'test_pulp_repository_rpm_repo_versions.keys' variable
that provides the versions. It starts empty, and may be set to the
latest versions in Ark via test-pulp-repo-version-query.yml, or a
specific set of versions via an extra variables file.

There still remain references to the group_vars versions for promotion,
which will be removed in a separate commit.