dev: specify package repository versions to promote manually
Currently, the set of package repository versions to promote in Ark is
defined via a group_vars file in the inventory.  This file was
overloaded, and its primary purpose was to define versions to sync from
Ark to test. In practice, we would rarely want to promote the latest
versions.

It has become clear that this repository is not the right place to store
'live' versions - that is stackhpc-kayobe-config.

This change avoids references to the group_vars versions when promoting,
and adds a new 'dev_pulp_distribution_rpm_promote_versions' variable
that provides the versions. It is empty by default, and should be set to
mapping of versions to promote in Ark via an extra variables file.

The group_vars version variables are no longer required and have been
removed.