Support filtering of package repositories and container images
Running the full workflows can take a long time, which may be
undesirable when we want to get changes only for a small number of
package repositories or container images (e.g. for a security or bug fix
release).

This change adds the ability to specify a filter to apply to the list of
package repositories and container images in the various playbooks. The
filter is also exposed in the Github actions workflows.