Fix updating Kayobe repository versions
Support for automatically updating the Kayobe Pulp repository versions
was broken by the addition of Rocky repositories, since these are not in
the Kayobe config.

This change fixes the issue by first querying the Pulp repository
versions in Kayobe config, then updating those based on the versions
queried from Ark.

This has the implication that all necessary version variables must now
exist in Kayobe config, and will not be automatically generated if
missing.