Support filtering of container images by regex
Running the full workflows can take a long time, which may be
undesirable when we want to get changes only for a small number of
container images (e.g. for a security or bug fix release).

This change adds the ability to specify a filter to apply to the list of
container images in the various playbooks. The filter is also exposed in
the Github actions workflows.

The filter takes the form of a whitespace-separated list of regular
expressions matching image names. This is the same format that is
accepted by the kolla-build tool.