source-repo-sync: Use the force with git add
This argument causes git to add ignored files to the repository. In some
cases there may be an entry in .gitignore that causes any added files to
be ignored. If no files are added, the workflow may fail with the
following:

  nothing to commit, working tree clean

This change fixes the issue by using git add . --force, which ignores
entries in .gitignore.