Fix container image promotion (hopefully)
This change modifies the way in which we promote container image tags.
Previously, the stackhpc-dev container repository was configured as a
remote, and the stackhpc repository synced with it, using include_tags
to limit to a specific tag.

Now, we use the "Add content" [1] API to add container tags from
stackhpc-dev to stackhpc. Currently we're still promoting a single tag
across all repositories, so a non-existent tag is not an error. This
will change in future.

[1] https://docs.pulpproject.org/pulp_container/restapi.html#tag/Repositories:-Container/operation/repositories_container_container_add