Promote multiple tags simultaneously based on Kayobe config
This modifies the container image promotion process to be similar to
package repo promotion. The tags for each image are stored in files in
StackHPC Kayobe configuration (one per base distro), and these are
queried to determine which tag to promote for each image.

This avoids the need to specify tags when promoting images, and will
allow for automatic tag promotion after merges to StackHPC Kayobe
Configuration.